Parsers: move result parser under generic name.
